uniform_int_distribution fix for msvc
Fixes: 
`error C2338: invalid template argument for uniform_int_distribution: N4659 29.6.1.1 [rand.req.genl]/1e requires one of short, int, long, long long, unsigned short, unsigned int, unsigned long, or unsigned long long `

+++ libs/io/src/CFileGZStreams_unittest.cpp
@@ -28,7 +28,8 @@ void generate_test_data(std::vector<uint8_t>& tst_data)
 {
 	const unsigned int random_seed = 123U;
 	std::mt19937 mersenne_engine{random_seed};
-	std::uniform_int_distribution<uint8_t> dist{0, 3};  // low entropy
+	// MSVC C++11 library enforces use of `unsigned short` as smallest type
+	std::uniform_int_distribution<unsigned short> dist{0, 3};  // low entropy
 	auto gen = [&dist, &mersenne_engine]() { return dist(mersenne_engine); };
 	// auto gen = []() { return 1; };
